+++ data/Autentication.js
@@ -7,22 +7,16 @@ export default class Autentication {
   }
 
   async login (type) {
-    await firebase.auth().setPersistence(firebase.auth.Auth.Persistence.LOCAL)
     let resultAuth = null
     if (type === 'Google') {
       resultAuth = await firebase.auth().signInWithPopup(this.provider)
-
-      return { name: resultAuth.user.displayName, photo: resultAuth.user.photoURL, email: resultAuth.user.email, phone: resultAuth.user.phoneNumber }
+      return resultAuth.user
     }
     return null
   }
 
-  async getUser () {
-    this.user = await firebase.auth().currentUser
-    if (this.user) {
-      return { name: this.user.displayName, photo: this.user.photoURL, email: this.user.email, phone: this.user.phoneNumber }
-    }
-    return null
+  async Logout () {
+    await firebase.auth().signOut()
   }
 
   isAuthenticated () {
